fix: upload model before training start
